New York Philharmonic Performance History

A tidy data version of the New York Philharmonic performance history data.

Usage

nyphil

Format

A tibble with 154,988 rows and 16 variables:

id

a character string denoting the GUID; to view program: archives.nyphil.org/index.php/artifact/GUID/fullview

program_id

a character string denoting the local NYP ID

orchestra

a character string denoting the full orchestra name

season

a character string denoting the season; defined as Sep 1 - Aug 31, displayed "1842-43"

event_type

a character string denoting the name of event time; see: http://archives.nyphil.org/index.php/help-performancehistory

location

a character string denoting the geographic location of concert; countries are identified by their current name. For example, even though the orchestra played in Czechoslovakia, it is now identified in the data as the Czech Republic

venue

a character string denoting the name of hall, theater, or building where the concert took place

date

a character string denoting the full ISO date used, but ignore TIME part (1842-12-07T05:00:00Z = Dec. 7, 1842)

time

a character string denoting the actual time of concert, e.g. "8:00PM"

id_2

a character string denoting the index number used to link soloist(s) and conductor(s) to a specific work on each of the concerts

composer_name

a character string denoting the composer name

work_title

a character string denoting the title of work; NYP short titles used

conductor_name

a character string denoting the conductor name

soloist_name

a character string denoting the soloist name (if multiple soloists on a single work, delimited by semicolon)

soloist_instrument

a character string denoting the soloist instrument name

soloist_roles

a character string denoting the soloist role: "S" means "Soloist"; "A" means "Assisting Artist" (if multiple soloists on a single work, delimited by semicolon)

Source

https://archives.nyphil.org/performancehistory